11239: Change dashboard to use refactored report code
