+++ app/views/welcome/_report_pane.html.erb
@@ -24,7 +24,8 @@
       </div>
     </div>
 
-    <div class="report-output">
+    <%# We don't use the modal but we have to use the partial with the modal for legacy reasons. %>
+    <div class="report-output-and-modal">
       <%= render("reports/output_and_modal") %>
     </div>
   <% end %>
